// backend/src/reservas/reserva.entity.d.ts
export declare class Reserva {
    id: number;
    peliculaId: number;
    peliculaTitulo: string;
    peliculaPortada: string;
    usuarioId: number;
    usuarioNombre: string;
    dia: string;
    horario: string;
    asientos: number;
    total: number;
    sala: string;
    estado: string;
    fecha: Date;
}